Adds a new --retries option to the patchbot
The default (0) is the same as the current behavior.  But for retries >= 1
the tests will be rerun up to 'retries' times with the --failed flag, to re-run
tests that failed on the initial run.

If the --failed run passes then the tests are considered passed, though a new
result TestsPassedOnRetry is used to indicate this.  This allows transient test
failures to be retried, but if this is happening constantly then it could still
indicate a real problem.