Fix bug when initial setup of the temporary SAGE_ROOT for an "unsafe" ticket fails
It fails to chdir out of the failed SAGE_ROOT, and to restore the
SAGE_ROOT environment variable to its original value.  Some of this
fix might be a bit redundant, but we make sure to take every opportunity
to set the current directory and SAGE_ROOT variable back to sane values
before trying to test other tickets.